--- services/backend/src/queue.json
@@ -1,1 +1,0 @@
-
