+++ services/backend/src/recommendation_system/recommendation_flow/candidate_generators/AlphaGenerator.py
@@ -8,7 +8,7 @@ from src.api.utils.auth_utils import get_user
 import json
 
 class AlphaGenerator(AbstractGenerator):
-    def get_content_ids(self, user_id, limit, offset, seed, starting_point):
+    def get_content_ids(self, user_id):
         print("user id from alpha generator:",user_id)
         # all images that have not been engaged by this user can be a candidate
         sql_statement_candidates = f"""
@@ -24,48 +24,3 @@ class AlphaGenerator(AbstractGenerator):
         # print('candidates from sql:',candidates)
         
         return candidates, None
-
-        # with open("/usr/src/app/src/queue.json", "r") as openfile:
-        #     try:
-        #         data = json.load(openfile)
-        #         print("data from queue.json",data,type(data))
-        #         queue = [i['queue'] for i in data if i['id']==user_id][0]
-        #     except: # if json file is empty, the case at initialization. or if it doesn't contain data for the current user
-        #         # data = []
-        #         queue = []
-
-        # print('queue',queue)
-
-        # print('starting_point',starting_point)
-
-        # if starting_point is None:
-        #     results = (
-        #         Content.query.with_entities(Content.id)
-        #     )
-        #     # print("results[:10]",results[:10])
-
-        #     # results = (
-        #     #     Content.query.with_entities(Content.id)
-        #     #     .order_by(func.random(seed))
-        #     #     .limit(limit)
-        #     #     .offset(offset)
-        #     #     .all()
-        #     # )
-        #     # print("results[:10]",results[:10])
-        #     print("if starting_point is None") #,list(map(lambda x: x[0], results)))
-        #     print("len(results)",len(list(map(lambda x: x[0], results))))
-        #     # print("results[:10]",list(map(lambda x: x[0], results))[:10])
-
-        #     results = list(map(lambda x: x[0], results))
-        #     results = [i for i in results if i not in queue]
-        #     print('results:',results)
-        #     # return list(map(lambda x: x[0], results)), None
-        #     return results, None
-
-        # elif starting_point.get("content_id", False):
-        #     content_ids, scores = ann_with_offset(
-        #         starting_point["content_id"], 0.9, limit, offset, return_distances=True
-        #     )
-        #     # print("elif starting_point.get("content_id", False)")
-        #     return content_ids, scores
-        # raise NotImplementedError("Need to provide a key we know about")
