--- services/backend/src/recommendation_system/recommendation_flow/controllers/AbstractController 2.py
@@ -1,3 +1,0 @@
-class AbstractController:
-    def get_content_ids(self, user_id, limit, offset, seed, starting_point):
-        pass
