+++ services/backend/src/recommendation_system/recommendation_flow/controllers/AlphaController.py
@@ -1,6 +1,3 @@
-# from src.recommendation_system.recommendation_flow.candidate_generators.RandomGenerator import (
-#     RandomGenerator,
-# )
 from src.recommendation_system.recommendation_flow.candidate_generators.AlphaGenerator import (
     AlphaGenerator,
 )
@@ -10,15 +7,9 @@ from src.recommendation_system.recommendation_flow.controllers.AbstractControlle
 from src.recommendation_system.recommendation_flow.filtering.AlphaFilter import (
     AlphaFilter,
 )
-# from src.recommendation_system.recommendation_flow.model_prediction.RandomModel import (
-#     RandomModel,
-# )
 from src.recommendation_system.recommendation_flow.model_prediction.AlphaModel import (
     AlphaModel,
 )
-# from src.recommendation_system.recommendation_flow.ranking.RandomRanker import (
-#     RandomRanker,
-# )
 from src.recommendation_system.recommendation_flow.ranking.AlphaRanker import (
     AlphaRanker,
 )
@@ -27,16 +18,11 @@ from src.recommendation_system.recommendation_flow.ranking.AlphaRanker import (
 class AlphaController(AbstractController):
     def get_content_ids(self, user_id, limit, offset, seed, starting_point):
         print('limit from AlphaController get_content_ids():',limit)
-        if seed <= 1:  # MySql seeds should be [0, # of rows] not [0, 1]
-            seed *= 1000000
-        candidates_limit = (
-            limit * 10 * 10
-        )  # 10% gets filtered out and take top 10% of rank
         candidates, scores = AlphaGenerator().get_content_ids(
-            user_id, candidates_limit, offset, seed, starting_point
+            user_id
         )
         filtered_candidates = AlphaFilter().filter_ids(
-            candidates, seed, starting_point
+            candidates
         )
         # predictions = RandomModel().predict_probabilities(
         predictions = AlphaModel().predict_probabilities(
@@ -51,7 +37,6 @@ class AlphaController(AbstractController):
             else {},
         )
         # print('predictions:',predictions)
-        # rank = RandomRanker().rank_ids(limit, predictions, seed, starting_point)
         rank = AlphaRanker().rank_ids(limit, predictions, seed, starting_point)
         print('rank',rank)
         return rank
