--- services/backend/src/recommendation_system/recommendation_flow/controllers/__init__ 2.py
@@ -1,9 +1,0 @@
-from src.recommendation_system.recommendation_flow.controllers.EngagementTimeController import (
-    EngagementTimeController,
-)
-from src.recommendation_system.recommendation_flow.controllers.RandomController import (
-    RandomController,
-)
-from src.recommendation_system.recommendation_flow.controllers.StaticController import (
-    StaticController,
-)
