--- services/backend/src/recommendation_system/recommendation_flow/filtering/AbstractFilter 2.py
@@ -1,3 +1,0 @@
-class AbstractFilter:
-    def filter_ids(self, content_ids, seed, starting_point):
-        pass
