--- services/backend/src/recommendation_system/recommendation_flow/model_prediction/AbstractModel 2.py
@@ -1,3 +1,0 @@
-class AbstractModel:
-    def predict_probabilities(self, content_ids, user_id, seed=None, **kwargs):
-        pass
