--- services/backend/src/recommendation_system/recommendation_flow/ranking/AbstractRanker 2.py
@@ -1,3 +1,0 @@
-class AbstractRanker:
-    def rank_ids(self, limit, probabilities, seed, starting_point):
-        pass
