--- services/backend/src/recommendation_system/recommendation_flow/ranking/DifeiRanker.py
@@ -1,30 +1,0 @@
-import heapq
-import random
-
-from .AbstractRanker import AbstractRanker
-
-
-class AlphaRanker(AbstractRanker):
-    def rank_ids(self, limit, probabilities, seed, starting_point):
-        k = int(limit * 0.9)
-        other = limit - k
-        top_limit = heapq.nlargest(limit, probabilities, key=lambda x: x["p_engage"])
-        top_k = top_limit[:k]
-        style = set([i['style'] for i in top_k if i['style'] == i['style']])
-        others = []
-        for i in probabilities:
-            if i['style'] not in style:
-                others.append(i)
-        top_other = heapq.nlargest(other, others, key=lambda x: x["explore"])
-        # if len(top_other) < other:
-        #     handle = top_other + top_limit[k:limit]
-        #     handle2 = sorted(set(handle), key=handle.index)
-        #     top_other = handle2[:other]
-
-        top_all = top_k + top_other
-        top_all_ids = list(map(lambda x: x["content_id"], top_all))
-        if seed:
-            random.seed(seed)
-
-        return random.sample(top_all_ids, len(top_all_ids))
-
